RA-1420 moved calculator to proper location.
