+++ src/Esfa.Vacancy.Application/Esfa.Vacancy.Application.csproj
@@ -65,6 +65,7 @@
     <Compile Include="Commands\CreateApprenticeship\CreateApprenticeshipRequest.cs" />
     <Compile Include="Commands\CreateApprenticeship\IMinimumWageSelector.cs" />
     <Compile Include="Commands\CreateApprenticeship\IWageTypeMapper.cs" />
+    <Compile Include="Commands\CreateApprenticeship\MinimumWageCalculator.cs" />
     <Compile Include="Commands\CreateApprenticeship\MinimumWageSelector.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\ContactDetailsValidator.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\CandidateCriteriaValidator.cs" />
