+++ src/Esfa.Vacancy.UnitTests/CreateApprenticeship/Application/GivenAMinimumWageCalculator/WhenCallingCalculate.cs
@@ -71,30 +71,4 @@ namespace Esfa.Vacancy.UnitTests.CreateApprenticeship.Application.GivenAMinimumW
                 .WithMessage($"{expectedMessage}\r\nParameter name: WageUnit\r\nActual value was 0.");
         }
     }
-
-    public class MinimumWageCalculator
-    {
-        private const decimal WeeksPerMonth = 4m;
-        private const decimal WeeksPerYear = 52m;
-        private const string MissingMinWageErrorMessage = "MinWage can't be null.";
-        private const string IncorrectWageTypeErrorMessage = "WageUnit must be either 'Weekly', 'Monthly' or 'Annually'.";
-
-        public decimal CalculateMinimumWage(CreateApprenticeshipRequest request)
-        {
-            if (!request.MinWage.HasValue)
-                throw new ArgumentOutOfRangeException(nameof(request.MinWage), request.MinWage, MissingMinWageErrorMessage);
-
-            switch (request.WageUnit)
-            {
-                case WageUnit.Weekly:
-                    return decimal.Divide(request.MinWage.Value, (decimal)request.HoursPerWeek);
-                case WageUnit.Monthly:
-                    return decimal.Divide(decimal.Divide(request.MinWage.Value, WeeksPerMonth), (decimal)request.HoursPerWeek);
-                case WageUnit.Annually:
-                    return decimal.Divide(decimal.Divide(request.MinWage.Value, WeeksPerYear), (decimal)request.HoursPerWeek);
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(request.WageUnit), request.WageUnit, IncorrectWageTypeErrorMessage);
-            }
-        }
-    }
 }
